Remove ModelTransformerContextBuilder and ModelTransformerContext interfaces
# Conflicts:
#	maven-api-impl/src/main/java/org/apache/maven/internal/impl/model/DefaultModelTransformerContextBuilder.java
#	maven-core/src/main/java/org/apache/maven/project/DefaultProjectBuilder.java
