+++ maven-api-impl/src/test/java/org/apache/maven/internal/impl/standalone/ApiRunner.java
@@ -158,7 +158,7 @@ public class ApiRunner {
 
         @Override
         public Path getRootDirectory() {
-            return null;
+            throw new IllegalStateException();
         }
 
         @Override
